refactor: 각 객체마다 사용하는 validator 생성
