+++ src/main/java/view/InputView.java
@@ -1,8 +1,5 @@
 package view;
 
-import laddervalidate.HeightValidator;
-import laddervalidate.PlayerNameValidator;
-
 import java.util.Arrays;
 import java.util.List;
 import java.util.Scanner;
@@ -13,25 +10,25 @@ public class InputView {
     private static final String DELIMITER = ",";
     private static final String HEIGHT_INPUT_REQUEST = "최대 사다리 높이는 몇 개인가요?";
 
-    private final PlayerNameValidator playerNameValidator = new PlayerNameValidator();
-    private final HeightValidator heightValidator = new HeightValidator();
+    private final InputPlayerNameValidator inputPlayerNameValidator = new InputPlayerNameValidator();
+    private final InputHeightValidator inputHeightValidator = new InputHeightValidator();
 
     public List<String> readUserNames() {
         System.out.println(NAME_INPUT_REQUEST);
         String inputUserNames = scanner.nextLine();
         List<String> playersNames = splitInputUserNames(inputUserNames);
-        playerNameValidator.checkDuplicatePlayers(playersNames);
-        playerNameValidator.checkPlayerNameLength(playersNames);
-        playerNameValidator.checkPlayerCount(playersNames);
+        inputPlayerNameValidator.checkDuplicatePlayers(playersNames);
+        inputPlayerNameValidator.checkPlayerNameLength(playersNames);
+        inputPlayerNameValidator.checkPlayerCount(playersNames);
         return playersNames;
     }
 
-    public String readHeight() {
+    public int readHeight() {
         System.out.println(HEIGHT_INPUT_REQUEST);
         String inputHeight = scanner.nextLine();
-        heightValidator.checkNumberMissMatch(inputHeight);
-        heightValidator.checkNegativeNumber(inputHeight);
-        return inputHeight;
+        inputHeightValidator.checkNumberMissMatch(inputHeight);
+        inputHeightValidator.checkNegativeNumber(inputHeight);
+        return Integer.parseInt(inputHeight);
     }
 
     private List<String> splitInputUserNames(String inputUserNames) {
